have added client-side authentication via session storage as an example of restricting access to routes -- user cannot access the todos nor welcome nor logout pages without logging in first
