+++ src/components/todo/AuthenticationService.js
@@ -1,0 +1,18 @@
+class AuthenticationService {
+    registerSuccessfulLogin(username, password) {
+        console.log("registerSuccessfulLogin");
+        sessionStorage.setItem("authenticatedUser", username);
+    }
+
+    logout() {
+        sessionStorage.removeItem("authenticatedUser");
+    }
+
+    isUserLoggedIn() {
+        let user = sessionStorage.getItem("authenticatedUser");
+        if (user === null) return false;
+        else return true;
+    }
+}
+
+export default new AuthenticationService()
