+++ src/components/todo/ErrorComponent.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+const ErrorComponent = () => {
+  return (
+    <div>
+      An Error Occrred. I don't know what to do. Contact support at
+      abcd-efgh-ijkl
+    </div>
+  );
+};
+
+export default ErrorComponent;
