+++ src/components/todo/FooterComponent.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+const FooterComponent = () => {
+  return (
+    <footer className="footer">
+      <span className="text-muted">
+        All Rights Reserved 2022 Blake Stickland
+      </span>
+    </footer>
+  );
+};
+
+export default FooterComponent;
