+++ src/components/todo/LogoutComponent.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+const LogoutComponent = () => {
+  return (
+    <div>
+      <h1>You are logged out</h1>
+      <div className="container">Thank you for using our application.</div>
+    </div>
+  );
+};
+
+export default LogoutComponent;
